requirements.txt

streamlit==1.31.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic==2.7.4
setuptools==70.0.0
wheel==0.43.0
textblob==0.18.0.post0
pandas==2.2.2
plotly==5.22.0
numpy==1.26.4  
typing-extensions==4.12.2
importlib-metadata==7.1.0